articles: fill in original dates where unambiguous
This patch adds a `date` field to the YAML frontmatter of each article.
The value is either a `YYYY-MM-DD` string representing the date on which
the first post in the article was originally authored, or `unknown` if
it is not immediately clear what that date should be. The date strings
are meant to represent local dates, presuming `America/New_York` locale.

I should be able to reverse-engineer at least a few of the unknown
dates, and will do so in follow-up patches.