articles: add date for "The Judgement Day"
I'm reasonably but not perfectly confident with this one. I can't find
this post in the College Confidential thread or anywhere else on the
web. But the post "I am in Remission!!!!!" is explicitly dated, and
seems to have been written the day after this one. This post refers to
judgment day as "tomorrow", and that post refers to judgment day as
"today"; both reference a rapture predicted for 2011-05-21 (although
this post refers to it as Sunday instead of Saturday; I'm chalking this
up as a typo, as neither the 2011 rapture nor the 2012 Mayan apocalypse
landed on a Sunday). Both posts also refer to having walked through the
worst-case scenarios to mentally prepare.