articles: add date for "A Beginning of a Journey: A Promise to Keep"
This post says that Jon will learn what is going on in five days, and
the post "Sharing the News with the Kids" is explicitly dated and
indicates that the news was shared on 2010-12-11. So, this post seems
likely to have been written five days earlier, on 2010-12-06.

This estimate is also consistent with "Interlude: Preparing for a
Crucial Conversation" having been written on 2010-12-09. Not only does
the timeline match up, but that post says that "last Monday I found out
that I have ovarian cancer"; that Monday was 2010-12-06. This post
*does* say that "[t]he biopsy results *yesterday* confirmed it is
cancer" (emphasis mine), but I think that it makes sense to interpret
this as meaning that the biopsy was "yesterday" and the results were
learned today, based on the processing described in this post.

This article was posted as post #766 in the College Confidential thread:
https://talk.collegeconfidential.com/t/my-diagnosis-of-advanced-cancer-how-to-help-my-kids/1013554/766
...but that post, from 2011-07-31, was actually posted long after it was
written ("I was going through some early writings [...]"), so an
estimate may be the best that we can do.