articles: move post date from filename to frontmatter
The dates in the filenames correspond to when the articles were posted
to Medium, not when they were written. This information is nice to keep
around, but is neither accurate (to original author date) or precise
(there are only four distinct values). Thus, this patch removes the
dates from the filenames and moves them to a new `postdate` field in the
frontmatter block.