articles: add date for "Hyeon Jeong Lee"
This one is more of a guess than most. This post was written by Amico
"over three years" after Hyeon died and is undated. However, there's a
separate College Confidential post by Amico with similar content and
similar wording (incl. verbatim phrase "from the day that she was
diagnosed with cancer" in the same context):
https://talk.collegeconfidential.com/t/the-writings-of-sunriseeast/1830125

Since this is the last undated article, I'm electing to use the date of
that College Confidential post as the date of this article. The timeline
lines up (posted 3.5 years after she died), and even if this date isn't
exactly correct, it's definitely the last article in the sort order,
which is one of the primary purposes of having these dates.