articles: link to College Confidential sources
This patch adds a `srcurl` field pointing to original sources of College
Confidential posts where appropriate. Almost all the articles have at
least one post to link to; I've linked to the first applicable post
where there are multiple.

The trick seems to be to do a within-thread search for phrases that
don't include names (which are often anonymized on College Confidential)
and to try searching from multiple parts of the article (since sometimes
only excerpts make it onto CC).