articles: standardize source types
In particular, various ad hoc forms of "this was posted to College
Confidential" are now standardized in the date line.

Format can change, of course, but this feels nicer than scattered
fragments of editorial metadata. I'm not certain how important this
information is, but Amico appeared to want to include it in the initial
compilation, so we can keep it around for now.